requirements.txt
pyspark==3.3.0
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.4 